+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+* Add link to source project to README
+* Update example project
+* Fix MaskedTextController bug where cursor goes to the beginning of the text
+
 ## 1.0.0
 
 * Initial release.